price to cents fix and larastan fixes

+++ app/Services/PriceService.php
@@ -6,12 +6,12 @@ class PriceService
 {
     //hello
 
-    public static function priceToEuro(int $priceInCents): float
+    public static function priceToEuro(int $priceInCents) : float
     {
         return $priceInCents / 100;
     }
 
-    public static function priceToCents(int $priceInEuro): int
+    public static function priceToCents(float $priceInEuro): int
     {
         return $priceInEuro * 100;
     }
